refactoring the variantinfo package
