+++ components/service/variantinfo/index.tsx
@@ -1,5 +1,6 @@
-export * from './variantInfo'
-export * from './myVariantInfo'
+export * from './variantInfoSources/alleleRegistryVariantInfo'
+export * from './variantInfoSources/gitDataHubVariantInfo'
+export * from './variantInfoSources/myVariantInfo'
 export * from './alleleSpecificEvidences'
 export * from './xQTLEvidences'
 export * from './externalRecords'
